full code base refactored to reuse Kafka functionality as much as possible (#44)
* Initial refactor

* Fixed build error

* With validator, principal builder & config

* replaced tab with spaces

* add configure interface for the validator

* Fix build error

* Cleaned up the code
