+++ azpubsub/.gitignore
@@ -1,0 +1,3 @@
+.cache-main
+.cache-tests
+/bin/
