--- azpubsub/api/src/main/scala/azpubsub/kafka/security/auth/InvalidTokenException.java
@@ -1,12 +1,0 @@
-package azpubsub.kafka.security.auth;
-
-public class InvalidTokenException extends Exception {
-
-    public InvalidTokenException(String error) {
-        super(error);
-    }
-
-    public InvalidTokenException(String error, Throwable c) {
-        super(error, c);
-    }
-}
