--- azpubsub/contextvalidator/src/test/java/azpubsub/contextvalidator/kafka/security/auth/mockPositiveSaslAuthenticationContextValidator.java
@@ -1,17 +1,0 @@
-package azpubsub.contextvalidator.kafka.security.auth;
-
-
-import azpubsub.kafka.security.authenticator.AzPubSubPrincipal;
-import azpubsub.kafka.security.authenticator.SaslAuthenticationContextValidator;
-
-import javax.security.sasl.SaslServer;
-
-public class mockPositiveSaslAuthenticationContextValidator implements SaslAuthenticationContextValidator {
-   public void configure(java.util.Map<String, ?> configs) {
-
-   }
-
-   public AzPubSubPrincipal authenticate(SaslServer saslServer) {
-      return new AzPubSubPrincipal("Token", "atesttoken");
-   }
-}
