--- azpubsub/contextvalidator/src/test/java/azpubsub/contextvalidator/kafka/security/auth/mockSslAuthenticationContextValidatorWithInvalidAuthentication.java
@@ -1,17 +1,0 @@
-package azpubsub.contextvalidator.kafka.security.auth;
-
-import azpubsub.kafka.security.authenticator.AzPubSubPrincipal;
-import azpubsub.kafka.security.authenticator.SslAuthenticationContextValidator;
-
-import javax.net.ssl.SSLSession;
-
-public class mockSslAuthenticationContextValidatorWithInvalidAuthentication implements SslAuthenticationContextValidator {
-    public void configure(java.util.Map<String, ?> configs) {
-
-    }
-
-    public AzPubSubPrincipal authenticate(SSLSession sslSession) {
-        return null;
-    }
-
-}
