+++ build.gradle
@@ -626,8 +626,6 @@ def pkgs = [
     'streams:test-utils',
     'tools',
     'azpubsub', 
-    'azpubsub:api', 
-    'azpubsub:contextvalidator'
 ] + connectPkgs
 
 /** Create one task per default Scala version */
@@ -868,10 +866,6 @@ project(':core') {
     from(project(':streams:examples').configurations.runtime) { into("libs/") }
     from(project(':azpubsub').jar) { into("libs") }
     from(project(':azpubsub').configurations.runtime) { into("libs") }
-    from(project(':azpubsub:api').jar) { into("libs") }
-    from(project(':azpubsub:api').configurations.runtime) { into("libs") }
-    from(project(':azpubsub:contextvalidator').jar) { into("libs") }
-    from(project(':azpubsub:contextvalidator').configurations.runtime) { into("libs") }
     duplicatesStrategy 'exclude'
   }
 
@@ -908,11 +902,7 @@ project(':core') {
     from(project(':streams:examples').configurations.runtime) { into("libs/") }
     from(project.file("$rootDir/repository")) { into "libs/" }
     from(project(':azpubsub').jar) { into("libs") }
-    from(project(':azpubsub').configurations.runtime) { into("libs") }
-    from(project(':azpubsub:api').jar) { into("libs") }
-    from(project(':azpubsub:api').configurations.runtime) { into("libs") exclude("zookeeper-3.4.14.jar") }
-    from(project(':azpubsub:contextvalidator').jar) { into("libs") }
-    from(project(':azpubsub:contextvalidator').configurations.runtime) { into("libs") exclude("zookeeper-3.4.14.jar") }
+    from(project(':azpubsub').configurations.runtime) { into("libs") exclude("zookeeper-3.4.14.jar") }
     duplicatesStrategy 'exclude'
     
     def copyDetails = []
@@ -978,10 +968,18 @@ project(':azpubsub') {
   }
 
   dependencies {
+    compile project(':core')
     compile libs.scalaLibrary
+    compile libs.jacksonDatabind
+    compile libs.jacksonJDK8Datatypes
+    compile libs.jacksonJaxrsJsonProvider
     compileOnly libs.log4j
 
     testCompile libs.junit
+    compile libs.easymock
+    compile libs.powermockJunit4
+    compile libs.powermockEasymock
+    compile libs.scalatest
   }
 
   jar {
@@ -991,99 +989,6 @@ project(':azpubsub') {
     }
   }
 
-  checkstyle {
-    configProperties = [importControlFile: "$rootDir/checkstyle/import-control-core.xml"]
-  }
-}
-
-project(':azpubsub:api') {
-  apply plugin: 'scala'
-  apply plugin: 'java'
-  archivesBaseName = "kafka-azpubsub-api"
-
-  sourceSets {
-    main {
-      java {
-        srcDir 'src/main/java'
-      }
-      scala {
-        srcDir 'src/main/scala'
-      }
-    }
-    test {
-      java {
-        srcDir 'src/test/java'
-      }
-      scala {
-        srcDir 'src/test/scala'
-      }
-    }
-  }
-
-  dependencies {
-    compile libs.scalaLibrary
-    compile libs.jacksonDatabind
-    compile libs.jacksonJDK8Datatypes
-    compile libs.jacksonJaxrsJsonProvider
-
-    testCompile libs.junit
-  }
-
-  jar {
-    manifest {
-      attributes('Implementation-Title': 'kafka-azpubsub-api',
-              'Implementation-Version': '1.0')
-    }
-  }
-
-  checkstyle {
-    configProperties = [importControlFile: "$rootDir/checkstyle/import-control-core.xml"]
-  }
-}
-
-project(':azpubsub:contextvalidator') {
-  apply plugin: 'scala'
-  apply plugin: 'java'
-  archivesBaseName = "kafka-azpubsub-contextvalidator"
-
-  sourceSets {
-    main {
-      java {
-        srcDir 'src/main/java'
-      }
-      scala {
-       srcDir 'src/main/scala'
-      }
-    }
-    test {
-      java {
-        srcDir 'src/test/java'
-      }
-      scala {
-        srcDir 'src/test/scala'
-      }
-    }
-  }
-
-  dependencies {
-    compile project(':core')
-    compile project(':azpubsub')
-    compile project(':azpubsub:api')
-      
-    compile libs.junit
-    compile libs.easymock
-    compile libs.powermockJunit4
-    compile libs.powermockEasymock
-    compile libs.scalatest
-  }
-
-  jar {
-    manifest {
-      attributes('Implementation-Title': 'kafka-azpubsub-contextvalidator',
-              'Implementation-Version': '1.0')
-    }
-  }
-    
   checkstyle {
     configProperties = [importControlFile: "$rootDir/checkstyle/import-control-core.xml"]
   }
